Añadido Error 404 para id inexistente en Firestore

+++ src/components/ItemDetailContainer.jsx
@@ -8,25 +8,35 @@ import { doc, getDoc } from 'firebase/firestore'
 
 function ItemDetailContainer () {
   const [item, setItem] = useState(null)
+  const [loading, setLoading] = useState(true)
+  const [error, setError] = useState(false)
 
   const { id } = useParams()
 
   const refDoc = doc(db, 'items', id)
 
   useEffect(() => {
-    getDoc(refDoc).then((snapshot) => {
-      const newItem = { id: snapshot.id, ...snapshot.data() }
-      setItem(newItem)
-    })
+    getDoc(refDoc)
+      .then((snapshot) => {
+        if (snapshot.exists()) {
+          const newItem = { id: snapshot.id, ...snapshot.data() }
+          setItem(newItem)
+        } else {
+          setError(true)
+        }
+        setLoading(false)
+      })
   }, [id])
 
-  return (
-    <>
-      {item === null
-        ? <h1 className='text-center m-5'>Loading...</h1>
-        : <ItemDetail item={item} />}
-    </>
-  )
+  if (error) {
+    return <h1 className='text-center m-5'>NOT FOUND 404</h1>
+  }
+
+  if (loading) {
+    return <h1 className='text-center m-5'>Loading...</h1>
+  }
+
+  return <ItemDetail item={item} />
 }
 
 export default ItemDetailContainer
